[Feature][WIP] Enemies (#12)
* update position of player

* working in animations player

* end loop sprites

* added files enemies + animatios player beta

* fixed pixel aprox and animations

* change position into float

* enemis basics + intento rotate

* limpieza

* Cleaning code

+++ assets/Resources/Collision.cpp
@@ -5,7 +5,7 @@ Collision::Collision(glm::mat4 *project, Collision::CollisionGroups group) {
     collisions = new glm::ivec4[1000];
     collisionGroup = group;
 
-    position = glm::ivec2(0, 0);
+    position = glm::vec2(0, 0);
 
     trianglesRenderer = TrianglesRenderer::createTriangleRender(project);
 }
@@ -14,20 +14,20 @@ Collision::Collision() {
     collidersSize = 0;
     collisions = new glm::ivec4[1000];
 
-    position = glm::ivec2(0, 0);
+    position = glm::vec2(0, 0);
 }
 
 Collision::~Collision() {
     collidersSize = 0;
 
-    position = glm::ivec2(0, 0);
+    position = glm::vec2(0, 0);
 }
 
 void Collision::setProjection(glm::mat4 *project) {
     trianglesRenderer = TrianglesRenderer::createTriangleRender(project);
 }
 
-void Collision::addCollider(const glm::ivec4 &boxCollider) {
+void Collision::addCollider(const glm::vec4 &boxCollider) {
     collisions[collidersSize] = boxCollider;
     collidersSize++;
 
@@ -44,16 +44,16 @@ void Collision::addCollider(const glm::ivec4 &boxCollider) {
     trianglesRenderer->addTriangle(triangle);
 }
 
-void Collision::removeCollider(const glm::ivec4 &boxCollider) {
+void Collision::removeCollider(const glm::vec4 &boxCollider) {
     // Unused
 }
 
-void Collision::changePositionAbsolute(const glm::ivec2 &pos) {
+void Collision::changePositionAbsolute(const glm::vec2 &pos) {
     position = pos;
     trianglesRenderer->moveHitBoxesRelative(pos);
 }
 
-void Collision::changePositionRelative(const glm::ivec2 &pos) {
+void Collision::changePositionRelative(const glm::vec2 &pos) {
     position += pos;
     trianglesRenderer->moveHitBoxesRelative(pos);
 }
