+++ assets/Resources/Collision.h
@@ -28,11 +28,11 @@ public:
 
     void setProjection(glm::mat4 *project);
 
-    void addCollider(const glm::ivec4 &boxCollider);
-    void removeCollider(const glm::ivec4 &boxCollider);
+    void addCollider(const glm::vec4 &boxCollider);
+    void removeCollider(const glm::vec4 &boxCollider);
 
-    void changePositionAbsolute(const glm::ivec2 &pos);
-    void changePositionRelative(const glm::ivec2 &pos);
+    void changePositionAbsolute(const glm::vec2 &pos);
+    void changePositionRelative(const glm::vec2 &pos);
 
 #ifdef SHOW_HIT_BOXES
     void showHitBox();
@@ -44,9 +44,9 @@ public:
     Collision::CollisionGroups collisionGroup;
 
     int collidersSize;
-    glm::ivec4 *collisions;
+    glm::vec4 *collisions;
 
-    glm::ivec2 position;
+    glm::vec2 position;
 
 private:
