+++ assets/Resources/CollisionSystem.cpp
@@ -50,7 +50,7 @@ bool CollisionSystem::searchForCollision(const Collision* a, const Collision* b)
     return false;
 }
 
-bool CollisionSystem::overlapVertical(const glm::ivec4 &a, const glm::ivec4 &b, const glm::ivec2 &posA, const glm::ivec2 &posB) {
+bool CollisionSystem::overlapVertical(const glm::vec4 &a, const glm::vec4 &b, const glm::vec2 &posA, const glm::vec2 &posB) {
     // if one rectangle is on left side of other
     if (a[0] + posA.x >= b[2] + posB.x || b[0] + posB.x >= a[2] + posA.x) {
         return false;
@@ -64,7 +64,7 @@ bool CollisionSystem::overlapVertical(const glm::ivec4 &a, const glm::ivec4 &b, 
     return true;
 }
 
-bool CollisionSystem::overlapHorizontal(const glm::ivec4 &a, const glm::ivec4 &b, const glm::ivec2 &posA, const glm::ivec2 &posB) {
+bool CollisionSystem::overlapHorizontal(const glm::vec4 &a, const glm::vec4 &b, const glm::vec2 &posA, const glm::vec2 &posB) {
     // if one rectangle is on left side of other
     if (a[0] + posA.x >= b[2] + posB.x || b[0] + posB.x >= a[2] + posA.x) {
         return false;
